fix(test): fix broken tests
Some tests failed because the used `alpine` service did start non-blocking which caused the container to stop immediately. Assertions expecting the container to run then failed.

The new structure is to use two blocking `nginx` images which serve a scenario of web and proxy.

For the `--abort-on-container-exit` flag tests a third - non-blocking by intend - service based on the `hello-world` image has been added.